fix: Specify ABI filters in NDK build
* Removes splicing logic
* Adds ABI filters for all supported ABIs

In older versions of ndk-build, downgrading the minimum SDK version and
splicing the aar files from builds targeting old and new ABIs was
required to make a working archive for all ABIs. Now that we are using
the new CMake options configuration, downgrades happen automatically for
ABIs requiring it.

armeabi stopped building automatically as it is deprecated, so now it
and other targets are specified in the abiFilter.

Fixes #5